Add format_temperature function
Add a new functionality to enable the design year or representative year application. Now the weather year can be used to certain time step (argument to_year). For example, if setting to_year = 2045, and the input temperature is for 2026, then the 2026 weather year will be used to calculate hdcd for time step 2045.